moved database package to source directory and created business logic package
