--- project/database/models/__init__.py
@@ -1,2 +1,0 @@
-from database.models.task import Category, Task
-from database.models.user import Profile
