+++ project/src/business_logic/providers.py
@@ -1,0 +1,47 @@
+from typing import NoReturn
+
+from fastapi import Depends
+from sqlalchemy.orm import Session
+
+from src.database.dependency import get_database_session
+from src.database.repositories import user, category, task
+from src.business_logic.services.auth import AuthService
+from src.business_logic.services.category import CategoryService
+from src.business_logic.services.task import TaskService
+
+
+def task_service_factory() -> NoReturn:
+    raise NotImplementedError
+
+
+def category_service_factory() -> NoReturn:
+    raise NotImplementedError
+
+
+def auth_service_factory() -> NoReturn:
+    raise NotImplementedError
+
+
+def task_service_provider(
+    session: Session = Depends(get_database_session),
+) -> TaskService:
+    task_repository = task.TaskRepository(session)
+    category_repository = category.CategoryRepository(session)
+    return TaskService(
+        task_repository=task_repository,
+        category_repository=category_repository,
+    )
+
+
+def category_service_provider(
+    session: Session = Depends(get_database_session),
+) -> CategoryService:
+    category_repository = category.CategoryRepository(session)
+    return CategoryService(category_repository=category_repository)
+
+
+def auth_service_provider(
+    session: Session = Depends(get_database_session),
+) -> AuthService:
+    user_repository = user.UserRepository(session)
+    return AuthService(repository=user_repository)
