+++ project/src/database/models/__init__.py
@@ -1,0 +1,2 @@
+from src.database.models.task import Category, Task
+from src.database.models.user import Profile
